Make `Value` ctor `explicit` and more concise. (#419)
* Add `operator==` for the `Value` class and its children.

These check whether or not the `Value`s have identical representation,
not whether they evaluate to the same value (which we cannot really know
at static time).

* Refactor to rely on `std::variant`'s `operator==`.

* Make `Value` ctor explicit and more concise.

Previously, the `Value` ctor was not explicit, which meant that the
various `Value` variants could implicitly become a `Value`. This is not
encouraged by the C++ style guide.

Further, this commit takes the various constructors defined for each
`variant` alternative and provides a single constructor that takes as
its argument the `variant` type. Because a given `variant` defines an
implicit conversion from each of its variants to itself, this is a more
concise way of accomplishing the same thing. To keep the restatement of
the `variant` type in sync between the field and the ctor, define
a `ValueVariant` `using` alias to use in both the field and the ctor
parameter.

* Make trivial children of `NamedValue` just expose its `==` with `using`.

* Rename `ValueVariants` to just `Variants`, since it is within `Value`.