Replace helper relations for specific-arity `operation`s with macros.
Previously, `operations.dl` declared a number of "helper" relations for
specific, common cases of `operation` such as `unaryOperation`,
`binaryOperation`, `assign`, etc. The problem with this is that it
requires each of those forms to be cross-populated to and from
`operation`, which increases toil.

This PR gets rid of the `unaryOperation` and `assign` helpers, which are
not currently used. It replaces the `binaryOperation` helper relation
with a `BINARY_OPERATION` helper macro, which just encodes a special
form for referring to the `operation` relation.