Replace helper relations for specific-arity `operation`s with macros. (#433)
Previously, `operations.dl` declared a number of "helper" relations for
specific, common cases of `operation` such as `unaryOperation`,
`binaryOperation`, `assign`, etc. The problem with this is that it
requires each of those forms to be cross-populated to and from
`operation`, which increases toil.

This PR gets rid of the `unaryOperation` and `assign` helpers, which are
not currently used. It replaces the `binaryOperation` helper relation
with a `BINARY_OPERATION` helper macro, which just encodes a special
form for referring to the `operation` relation.

Closes #433

COPYBARA_INTEGRATE_REVIEW=https://github.com/google-research/raksha/pull/433 from google-research:unary_binary_operations_as_macros@winterrowd 404c45775568835be3c74d0b214205848de26ae0
PiperOrigin-RevId: 433285522