Add marker dependencies to indicate targets that need protobufs.
Bazel implicitly adds the protobuf library as an implicit dependency on
every target. This makes it hard to perform an automatic import into
Google's internal repo effectively. This PR adds a `protobuf` or
`message_differencer` dependency to every target that needs protocol
buffers that is not some kind of proto library already. This allows us
to automatically tranform these targets to the internal Google
equivalent in a more robust fashion.