Decode `TagTransform` proto.
The `TagTransform` proto indicates that some policy may perform a cast
upon the tag state. The details of that cast are considered a policy
aspect, and are not handled in this structure, which is considered
a dataflow graph aspect structure. Instead, the `TagTransform` carries
a policy name which can be used for linking to this rule in the dataflow
analysis.