Make input/output CLI more normal (#463)
Make the new CLI:
```
./authorization_logic <input_file_path> -o <output_file_path>
```
instead of the fully ridiculous
```
./authorization_logic <filename> <input_file_path_except_filename> <output_file_path_except_filename> (and .dl gets magically added to output filename)
```

This should also make the bazel rules for running this thing a bit easier to write because bazel better handles programs where the complete paths of generated outputs are verbatim in the command invocation somewhere.

A followup PR will allow accepting multiple files as inputs (without checking cryptographic signatures) from the CLI which we need for binary transparency.

Closes #463

PiperOrigin-RevId: 436753229