Allow the ID on SQL IR expressions to be optional.
Previously, we required every expression to be given a unique ID. This
was so `TagTransform`s would always be able to point uniquely to their
precondition expressions. However, during the integration with the SQL
verifier, I realized that assigning unique IDs to every expression was
harder than just assigning unique IDs to those expressions that need to
be pointed to by `TagTransform`s.

This PR makes Raksha handle expressions with a missing id (or,
equivalently in Protobuf, explicitly given the id 0) by just not adding
them to the map from IDs to expressions. Otherwise, they are
deserialized exactly the same as expressions with an ID.