Change the `TagTransform` preconditions to be a map from name to ID. (#489)
While integrating with the SQL verifier that we wish to integrate with
Raksha, I noticed that it associates each of the preconditions on its
tag transforms with a name in an unordered map. This is an awkward fit
with the way we had originally written the `TagTransform`, which would
associate its preconditions with the preconditions of a policy rule by
an order imposed on the preconditions. For one, it would require me to
repeatedly sort the preconditions on the verifier side when building the
protos, repeatedly performing an O(n * log(n)) operation. For another,
the names would preserve debugging information in Datalog and probably
produce better results if a mechansim like Souffle's `explain` were
used.

For these reasons, this PR modifies the `TagTransform` preconditions to
be a map from precondition name to the ID of the AST node that will be
associated with that precondition. These names are used as the names of
the `input`s to the `TagTransform`, and can be matched by the policy
datalog produced from the SQL verifier's rules.

Closes #489

COPYBARA_INTEGRATE_REVIEW=https://github.com/google-research/raksha/pull/489 from google-research:tag_transform_precondition_map@winterrowd 4201187ed58f6e322548035f778bf3cebe70eb0a
PiperOrigin-RevId: 437293023