Make `DecodeExpressionArena` have a no-context overload.
`DecodeExpressionArena`, up til now, has taken a `DecoderContext` as one
of its arguments. You could imagine this being very useful for fancy
scenarios, such as having multiple pieces of a large program containing
a SQL query relying on a common context. However, it is a bit annoying
for the simple case we would like to solve right now, in which we want
to decode the IR for a single SQL query for the SQL verifier
integration. For this case, it is easier to not pass a context, which
should cause a fresh context to be constructed and used for decoding.
This commit adds an overload with that behavior.