Create C++ types for Souffle values which can be printed to Datalog.
This change adds C++ types which directly map to Souffle's types. These
types use templates to allow a user to create C++ types that correspond
directly to Souffle types and which the C++ compiler can type check at
static time. Because some of the types that we create are rather
complex, involving both ADTs and records, this should be useful to catch
errors before having to run the Souffle compiler.