Add C++ types to express Souffle facts using C++ Souffle value types.
Adds a `Fact` class that allows expressing a Souffle fact as a C++ type,
with the layout expressed in terms of the C++ Souffle value types
provided in an earlier PR. Using these Souffle value types in templates
to describe the layout allows the C++ compiler to check our work for us.
Provides methods for printing `Fact`s out as datalog.