Add a representation of the `IsOperation` fact in C++.
Add a C++ representation of the Datalog fact `IsOperation`, used by
Raksha to get the list of `Operation`s that the analysis should concern
itself with. Due to the generic Datalog types already created, creating
this representation requires no new executable code, just new type
definitions. Created some tests showing some plausible SQL types that we
wish to represent being expressed in this system.

While trying to get this to build, I set off the warning
`-Wsubobject-linkage`. I think this is due to the extensive use of
`using` aliases in the headers created. It looks from this stackoverflow
question
(https://stackoverflow.com/questions/41167119/how-to-fix-a-wsubobject-linkage-warning)
like that warning may be overly strict on `using` aliases, so disabling
it in the `.bazelrc` file.