Add a boolean conversion to `intrusive_ptr`.
This change adds a boolean conversion to `intrusive_ptr`. While type
conversions are frowned upon by the Google style guide, boolean
conversions of this form are highly idiomatic for C++ smart pointers,
and are defined on both `std::unique_ptr` and `std::shared_ptr`. This
method allows more syntactically-fluid use of the `GetIf` function on
`Attribute`, as with it we can write code like so:

```
if (shared_ptr<Int64Attribute> int64_attr =
  attr.GetIf<Int64Attribute>()) {
...
}
```

with the conversion to `bool` allowing for the testing of the introduced
`int64_attr` local.