Remove kRelationName from template parameters on InputRelationFact.
Using constexpr char[] globals in template parameters is a trap. There
is no way to use a pointer to a global with linkage as a template
argument; any string that you manage to place in such a template
parameter must not have linkage. This means that it will probably have
surprising results once you attempt to use this type across TU
boundaries. Just getting rid of it and moving `InputRelationFact` to
being an abstract class with a `GetRelationName` method that needs to be
filled in by subclasses.