Create a new `refct_ptr`, an intrusive_ptr that brings its own counter.
The `intrusive_ptr` smart pointer expects its pointed-to type to have
a counter upon it. This means that the pointed-to object needs to derive
from `RefCounted` or duplicate its interface to allow `intrusive_ptr` to
count correctly. But what if you want a shared smart pointer and you
don't want to worry about inheriting from a counter object?

That is what the new `refct_ptr` tries to do. Rather than requiring its
payload to bring its own counter, at time of allocation, it composes
that payload with a `WithRefCount` object that provides the `RefCounted`
superclass. This allows `refct_ptr` to point to any object with few
additional restrictions at the cost of an extra `uint64_t` in size on
the pointed-to payload.