Add ctor to `IRContext` allowing creating with initial contents.
Prior to this commit, all `Operator`s and `Storage`s only got into an
`IRContext` by being passed into their respective `Register` function.
While this makes sense for many circumstances, it makes it harder to
write certain tests, as objects that need an `IRContext` cannot be
constructed in the static scope without global mutable state. In
addition, it could make sense to have an `IRContext` with a known,
initial set of `Operator`s and `Storage`s: if we know an exact set of
allowed `Operation`s in advance, why not just build the `IRContext` with
them in there from the start?

This commit adds a constructor to `IRContext` that takes a `std::vector`
of initial `Operation`s and `Storage`s to seed the `IRContext` with. The
constructor then just calls `Register` for each of those items. This
commit also uses this new constructor in a test where it can make the
test more concise.