Add constructors for the SQL operations.
This commit adds constructors to the SQL operations that allow these
operations to be built more easily at the static scope for tests. The
existing `Create` methods are defined in terms of these constructors.

Note for reviewers: see comment in MergeOp, would be interested in
feedback on how to make that better.