Calculate positional indices of operands to operations.
Because lists in Souffle are just recursive linked-list structs, it is
not very natural to look up elements by index. For the tag transform
operation from the SQL verifier, however, we need to do exactly that.
This commit introduces `operationHasOperandAtIndex`, which enriches the
old `operationToOperands` relation by adding the position the operand
holds in the operation's `operandList`.