Add "sql_output" operation and report if any tags reach this op.
The SQL verifier we wish to support wants to emit an error if any
confidential tag reaches any top-level output. To handle this, this
commit adds a "sql.sql_output" operation that the analysis checks for
any confidentiality tag and outputs the result ("PASS" or "FAIL")
alongside the operation into a `sqlOutputResult` relation.

In addition, this change creates an interface that takes as inputs the
`isOperation` and `isSqlPolicyRule` relations and outputs the
`sqlOutputResult` relation. This interface is probably sufficient for
our first cut at the SQL verifier.