Wrap the top `Value` from an arena with a new `SqlOutputOp`.
In a previous commit, we introduced the concept of a "sql.sql_output"
operator in datalog, which would cause a failure if any confidentiality
tag made it to the access path wrapped by that operation. This
introduces the code to use that concept in the IR translator by wrapping
the top level value with an operation having operator `SqlOutputOp`.