Add a top-level driver to tie the Raksha SQL analysis together.
This commit adds a top level driver that ties together the pieces of the
SQL analysis, causing a dataflow graph proto and a string containing
datalog `isSqlPolicyRule` facts to be turned into a boolean, indicating
whether the policy was violated or not.

This change also refactors the creation of temporary directories by
creating a function in `utils` for doing it the right way that can be
called in multiple locations.