Add a `raksha_cc_proto_library`, which will be used to simplify import.
The versions of `cc_proto_library` inside and outside of Google are
slightly incompatible in the attributes that they take. This commit adds
a new `raksha_cc_proto_library` which will have an internal and external
version to help ease the translation between the two versions. It is
intentionally not yet used; we will finish setting up the translation on
the Google-internal side, as that's where most of the work needs to
occur.