Add a proto representation of the Raksha IR. (#562)
This PR adds a representation that directly captures the Raksha IR.
Unlike our previous attempt at creating a SQL-specific IR proto, this
proto maps directly to the Raksha IR structures, and does not decide on
the set of allowed operations in advance. Instead, it defers to the
frontend to provide the operations that that frontend is interested in.

In an attempt to future-proof the IR, this introduces one structure for
which there is no analog in the current IR: an `IrTranslationUnit`
message. The `IrTranslationUnit` contains a single top-level module
plus some additional fields to annotate that module (currently, just the
name of the frontend from which the IR originated). While not extremely
useful at the moment, this top-level structure gives us a place to put
things relevant to the analysis on the whole TU rather than any
individual structure. These things may include analysis tuning or
checker-selecting options, the commit hash or version string of the
frontend, the person or entity requesting the analysis, etc.

Closes #562

COPYBARA_INTEGRATE_REVIEW=https://github.com/google-research/raksha/pull/562 from google-research:generic_ir_protos@winterrowd c9145913284af49c296e4cc79fc18808a2687a44
PiperOrigin-RevId: 451010740