Add list of `Operator`s to IR TU message.
This PR adds an `Operator` message and make the IRTranslationUnit
message contain a list of `Operator`s. We expect that eventually,
`Operator`s will contain more information than just the name they
currently contain, and thus it will make sense to have an IR TU
declare/define the list of `Operator`s that it will be using. In
addition, the `IRContext` object currently requires that `Operator`s be
added to the context before any `Operation` using that `Operator` is
built. Adding the list of possible `Operator`s to the TU allows those
`Operator`s to be populated before a decoder attempts to decode any
particular `Operation`.