Add "magic" tag xforms to change itag propagation behavior on merges.
While integrating Raksha into the SQL verifier that we are targeting, we
discovered that the SQL protos that we created are factored wrongly, and
in the future we would like frontends to target protos that mirror the
structure of the IR directly. However, to complete the integration of
Raksha and the SQL frontend, we discovered that we need to support merge
operations that have different ways of propagating integrity tags. To do
this the right way would involve modifying the SQL IR protos, which
would be a lot of work on something we now know we want to throw away.

This commit introduces a hack to get us the rest of the way to
supporting the SQL verifier using the current SQL IR protos without
incurring the code changes needed to update the SQL IR protos. We
introduce some specially-named TagTransform operations that modify the
integrity tag propagation behavior of the merge operations that they
wrap. While not ideal, this allows us to get the behavior we want with
the current SQL IR structure.