Eliminate duplicate lists of dl files with a shared .bzl file.
In various places, we list the same (or very similar) lists of dl files. This is largely due to the non-modularity of Souffle's building of dl files. This change attempts to reduce the error-prone-ness of this duplication by putting some of the common lists of dl files into a shared `.bzl` file, co-located with those `.dl` files in the src/analysis/souffle directory.

PiperOrigin-RevId: 465392391