Support additional characters in constant names (#593)
To support the transparent release process, we need to allow
constant names to include: -, +, /, =, spaces, and newlines. This is
needed to include the contents of public key files and URIs in
variables. The test included in `.../src/tests/test_hyphen_id.rs` and
`.../src/test_inputs/hyphenIDs` illustrates the use-case.

As a side-effect, the lex rules for variables (identifiers not in
quotes) and constants (identifiers in quotes) are more cleanly separate,
which should be a strict improvement in the grammar.