Simplify the DP parameter checker interface.
The initial version of the DP parameter interface indicated the kind of parameter being passed with both a name and the ADT type. This is redundant, and made producing these facts more verbose and harder to test. This change makes the DP parameter interface use only the ADT to indicate the type of DP parameter.

PiperOrigin-RevId: 468273716