Connect the IR translator to the DP analysis.
This change connects the IR translator to the DP analysis. It introduces a new kind of policy object, indicating global limits on DP parameters, and shows this policy being used to cause the analysis to return a pass and a fail on the same IR graph depending upon those parameters.

PiperOrigin-RevId: 468717689