Remove default integrity tag propagation.
This change removes Raksha's default propagation of integrity tags. Integrity tags propagate under very limited circumstances, and it is hard (and possibly performance-intensive) to perform this default propagation. Eliminating propogation of integrity tags and forcing the implementers of operations to tell us how they would like their integrity tags to be propogated has almost the same effect with potentially lower complexity. This is just an experiment for now.

PiperOrigin-RevId: 473355774