Make Raksha build with 64-bit Souffle.
Previously, we had talked about the fact that using Souffle in 64 bit mode would be a good idea. We have even structured IR nodes expecting that Souffle would be in 64 bit mode. Unfortunately, it seems that we never actually set up Souffle to compile in 64 bit mode or we somehow regressed back to the default 32 bit mode. This CL makes us build Souffle in 64 bit mode.

PiperOrigin-RevId: 473792486