Create a tool to convert Raksha IR files into Souffle isOperation facts.
This CL introduces a new command line tool, `ir_to_operation_facts`, that converts a Raksha IR file
into a group of `isOperation` facts, suitable for sending into souffle via the `isOperation` input
relation. This is useful for writing unit tests, as you can write a dataflow graph in the Raksha IR
rather than hand-writing isOperation facts, which are hard to read and a pain to write.

PiperOrigin-RevId: 474324575