Eliminate SsaNames' tracking of IDs for `Operation`s, rely only on Values.
Prior to this commit, the IRPrinter suffered from a nasty bug where its SSA numbering failed to cohere. This was due to the fact that we were counting the SSAs for `Operation`s in two different places: in a map keyed on `Operation`s for the result values, but on a map keyed on `Value`s for the inputs. This commit fixes that by eliminating the `Operation` map, relying only on `Value`s for providing `SSA` numbers. Thinking about this intuitively, this makes sense because SSA means "each name is statically assigned a single *value*"; `Operation`s, on their own, are not `Value`s!

Fixes #690

PiperOrigin-RevId: 476199472