Deprecate relations to not declare
In the very early days of Raksha when we were connecting
the rust implementation of authorizaiton logic to a datalog backend
of Arcs, we needed to skip some relation declarations as a hack
to get this to work. We have implemented a similar feature in the C++
implementation of authorization logic, but we do not need this. Also,
once we have integration tests for the C++ implementation of auth logic,
skipping some of these declarations will now cause the compiler not to
work. To that end, this PR deprecates that small bit of code.

No additional tests are added, but notably when this (dead) code
is deleted, the existing test that run on PRs should pass.