Replace absl::string_view with std::string_view.
We are currently trying to give ourselves the option to use Abseil when the performance gains of that library are worth it but give ourselves the option of avoiding it when it is not available. `absl::string_view` is an interesting case: Google does not use it for [mysterious compatibility reasons](https://abseil.io/tips/1). Within our own repo, we are unlikely to care about those and `absl::string_view` offers no performance gain over `std::string_view`, so just switch to `std::string_view`.

PiperOrigin-RevId: 485412041