Refactor printing of values to remove circular dependencies.
Prior to this change, `Value` contained a `ToString` method taking an `SsaNames`. This is getting very close to a circular dependency, as `SsaNames` includes `value.h`, which it needs because it takes `Value` by value in some of its methods. The only way this circular dependency was avoided was by putting the implementation of `ToString` into `ir_printer.h`, which is a bit of an unusual pattern.

While that placement of `ToString`'s implementation fixed that one instance of the problem, it still leaves the repo in a state where circular dependencies are easy to make. Adding `Value`s to the body of `Storage` (as I'll do in a follow-on commit) causes this circular dependency to pop up again.

This change undoes the tangle by removing `ToString` and placing it in a separate helper class, which can depend upon both `value.h` and `ssa_names.h` without circularity.

PiperOrigin-RevId: 491393417