Add `input_values` to `Storage`.
As part of integrating `Storage`s into the IR graph in a more full-featured way, this change adds a set of `input_values` to the `Storage` object, similar to how an `Operation` has a list of `inputs`. The difference in container choice is a consequence of the fact that `Operation`s take a tuple of inputs where order matters, while `Storage`s just keep track of their inputs in an order-agnostic fashion.

Making this change had surprisingly wide-ranging effects upon `#include`s and build rules. Causing `Storage`s to take a list of input values caused a circular dependency between the `value` library and the `storage` library. I resolved this by moving the functionality for printing `StoredValue`s into `value_string_converter.h`.

In addition, this change caused me to replace an inclusion of `storage.h` with a forward declaration where the full declaration was not needed. That exposed a large number of cases where `Storage` was included via a transitive header inclusion, which then needed to be addressed.

PiperOrigin-RevId: 491461317