Remove `impl_module_` from `Operation`.
In earlier drafts of the Raksha IR, we imagined that an `Operation` may contain a `Module` describing its implementation. We never got around to using this field in any significant way. Lately, we have decided that `Module`s may be the natural place to put `Storage`s, which would make this nesting nonsensical: `Operation`s are supposed to be stateless, so having a stateful structure in their implementation would not work. This change removes this member from `Operation` and resolves this semantic conflict.

PiperOrigin-RevId: 491462765