Refactor the IR parser to use a visitor structure.
Previously, the IR parser walked the Antlr parse tree without using Antlr's visitor or listener structures. While this parser was done correctly, it was brittle against changes in the grammar. Because the structure was not factored according to the "joints" of the parse tree's grammatical elements, it was hard to update the code walking the parse tree correctly in response to changes in the grammar.

This change refactors the parser to use Antlr's visitor class. This allows more readily finding where the code must change to accomodate grammatical changes.

This refactoring shook out a few bugs in the process:

1. Previously, string literals would only allow those characters that could be used in an identifier. Now they allow all characters except `"` (escaping characters is still TODO).

2. Previously, you could get away with assigning a temporary multiple times with the same value. While benign, this is a violation of the "single" part of the "single static assignment" scheme we are using for temporaries.

3. Previously, the grammar did not necessarily have to match the entire string it was given. Now the top level `IrProgram` rule must match `EOF`, enforcing a full match.

New quirk: if your file ends with a `//` comment, it must be newline-terminated, otherwise the lexer gets unhappy (because it expects `//` comments to end in a newline).

PiperOrigin-RevId: 491954284