Add `Storage`s to `Module`.
This change makes the `Module` structure contain `Storage`s, which causes it to be the place where mutable state lives. This is in contrast to `Operation`s, which are stateless. We will hook up these creation and fetching methods to the parser(s) and the solver to instantiate `Storage`s and hook them into the IR graph.

PiperOrigin-RevId: 491971070