Refactor to create Souffle programs directly rather than use factory.
Souffle's interface uses a pattern wherein every generated analysis is associated with some name in a global factory object. The preferred use of Souffle is to ask the program factory for the name of your analysis, which it will then look up in the map and give to you. This works, but encourages all analysis libraries to be linked to the one place where you call Souffle, whether or not you use them.

Despite the pattern Souffle has put in place, there is another way. Souffle also generates a `newInstance` method that returns an instance of the `SouffleProgram`. Souffle doesn't give us a generated header that is convenient to get this function, but we can "guess" the name of this function to have the linker get it for us at link time. This reduces the dynamism in getting Souffle analysis objects.

Statically naming the Souffle `newInstance` method gives us build-time errors if we got the name wrong (unlike mysterious runtime errors using the strings). It encourages linking the Souffle program with the policy that will use it, which produces a less bloated program structure.

The big reason I wanted to do this, though, is to see whether this additional bit of static information allows the linker to be more intelligent about what functions are kept and what functions are removed during link time optimizations. There is some indication that this helps a surprising amount.

PiperOrigin-RevId: 492874727