Remove `IsConst` from `IRTraversingVisitor`, which can't actually use it.
The `IRTraversingVisitor`, like its parent class the `IRVisitor`, had a `IsConst` template parameter which determined whether the AST nodes it visits should be treated as `const` or not. While this could make sense for the `IRVisitor` superclass, which does not indicate how it will visit these IR structures, it does make sense for the `IRTraversingVisitor`, which walks over collections which are const references, making a non-const traversal impossible.

PiperOrigin-RevId: 493792481