Introduce an `IndexedHashMap`, which has constant time lookup and stable iteration order.
This change introduces an `IndexedHashMap`, which has O(1) lookup but stable iteration order. Under the hood, this is just a combination of a `absl::flat_hash_map` from key to index and a `std::vector` containing the values. This can't really use a standard C++ interface due to the key and the value not being co-located, but we can provide a sensible alternative map interface for it.

PiperOrigin-RevId: 501124236